requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dataclasses-json==0.6.7
distro==1.9.0
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
filetype==1.2.0
frozenlist==1.8.0
google-auth==2.49.1
google-genai==1.67.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.12
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.19
langchain-google-genai==4.2.1
langchain-mcp-adapters==0.2.1
langchain-openai==1.1.11
langchain-text-splitters==1.1.1
langgraph==1.1.2
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.17
marshmallow==3.26.2
mcp==1.26.0
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.3
ollama==0.6.1
openai==2.28.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
propcache==0.4.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyGithub==2.8.1
PyJWT==2.12.1
PyNaCl==1.6.2
python-dotenv==1.2.2
python-multipart==0.0.22
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
sniffio==1.3.1
SQLAlchemy==2.0.48
sqlparse==0.5.5
sse-starlette==3.3.2
starlette==0.52.1
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
